Revert "Fix usage of Emacs on Heroku"
This reverts commit 1b3bbb29540dca32837891cc8b884124e3bcabb8.

It turns out that setting the EMACS environment variable doesn't let
you use a different Emacs binary for the GNU ELPA build system. In
fact, there's no way to do that. We have to establish a symlink. Which
is stupid. (Although the fact that Ubuntu's `emacs` package doesn't
install an `emacs` binary is equally stupid.)